+++ src/domains/portfolio/components/ImportWallet/ImportWallet.WIF.test.tsx
@@ -1,4 +1,4 @@
-import { Contracts } from "@ardenthq/sdk-profiles";
+import { Contracts } from "../../../../../../platform-sdk/packages/profiles/source/helpers";
 import userEvent from "@testing-library/user-event";
 import { createHashHistory } from "history";
 import React from "react";
@@ -6,8 +6,8 @@ import { FormProvider, useForm } from "react-hook-form";
 import { Route } from "react-router-dom";
 
 import { MethodStep } from "./MethodStep";
-import { EnvironmentProvider } from "@/app/contexts";
-import { OptionsValue } from "@/domains/wallet/hooks/use-import-options";
+import { EnvironmentProvider } from "../../../../app/contexts";
+import { OptionsValue } from "../../../wallet/hooks/use-import-options";
 import {
 	env,
 	getDefaultProfileId,
@@ -16,8 +16,8 @@ import {
 	waitFor,
 	mockProfileWithPublicAndTestNetworks,
 	act,
-} from "@/utils/testing-library";
-import * as usePortfolio from "@/domains/portfolio/hooks/use-portfolio";
+} from "../../../../utils/testing-library";
+import * as usePortfolio from "../../hooks/use-portfolio";
 
 let profile: Contracts.IProfile;
 const fixtureProfileId = getDefaultProfileId();
