+++ src/domains/portfolio/components/ImportWallet/ImportWallet.test.tsx
@@ -1,4 +1,4 @@
-import { Contracts } from "@ardenthq/sdk-profiles";
+import { Contracts } from "../../../../../../platform-sdk/packages/profiles/source/helpers";
 import userEvent from "@testing-library/user-event";
 import { createHashHistory } from "history";
 import React, { act } from "react";
@@ -8,9 +8,9 @@ import { Route } from "react-router-dom";
 import { ImportWallet } from "./ImportWallet";
 import { MethodStep } from "./MethodStep";
 import { SuccessStep } from "./SuccessStep";
-import { EnvironmentProvider } from "@/app/contexts";
-import { translations as commonTranslations } from "@/app/i18n/common/i18n";
-import { OptionsValue } from "@/domains/wallet/hooks/use-import-options";
+import { EnvironmentProvider } from "../../../../app/contexts";
+import { translations as commonTranslations } from "../../../../app/i18n/common/i18n";
+import { OptionsValue } from "../../../wallet/hooks/use-import-options";
 import {
 	env,
 	getDefaultProfileId,
@@ -21,8 +21,8 @@ import {
 	mockNanoXTransport,
 	mockProfileWithPublicAndTestNetworks,
 	mockProfileWithOnlyPublicNetworks,
-} from "@/utils/testing-library";
-import * as usePortfolio from "@/domains/portfolio/hooks/use-portfolio";
+} from "../../../../utils/testing-library";
+import * as usePortfolio from "../../hooks/use-portfolio";
 
 let profile: Contracts.IProfile;
 const fixtureProfileId = getDefaultProfileId();
