+++ src/domains/portfolio/components/ImportWallet/ImportWallet.tsx
@@ -1,4 +1,4 @@
-import { Contracts } from "@ardenthq/sdk-profiles";
+import { Contracts } from "../../../../../../platform-sdk/packages/profiles/source/helpers";
 import React, { useEffect, useMemo, useState } from "react";
 import { useForm } from "react-hook-form";
 import { useTranslation } from "react-i18next";
