+++ src/domains/portfolio/components/ImportWallet/Ledger/LedgerConnectionStep.tsx
@@ -1,4 +1,4 @@
-import { Networks } from "@ardenthq/sdk";
+import { Networks } from "../../../../../../../platform-sdk/packages/sdk";
 import React, { useEffect, useState } from "react";
 import { useFormContext } from "react-hook-form";
 import { useTranslation } from "react-i18next";
