+++ src/domains/portfolio/components/ImportWallet/Ledger/LedgerTabs.Cancel.test.tsx
@@ -1,14 +1,14 @@
-import { Contracts } from "@ardenthq/sdk-profiles";
+import { Contracts } from "../../../../../../../platform-sdk/packages/profiles/source/helpers";
 import userEvent from "@testing-library/user-event";
 import React, { useEffect } from "react";
 import { FormProvider, useForm } from "react-hook-form";
 import { Route } from "react-router-dom";
 
 import { LedgerTabs } from "./LedgerTabs";
-import { minVersionList } from "@/app/contexts";
-import { env, getDefaultProfileId, render, screen, waitFor, mockNanoXTransport } from "@/utils/testing-library";
-import { useLedgerContext } from "@/app/contexts/Ledger/Ledger";
-import { server, requestMock, requestMockOnce } from "@/tests/mocks/server";
+import { minVersionList } from "../../../../../app/contexts";
+import { env, getDefaultProfileId, render, screen, waitFor, mockNanoXTransport } from "../../../../../utils/testing-library";
+import { useLedgerContext } from "../../../../../app/contexts/Ledger/Ledger";
+import { server, requestMock, requestMockOnce } from "../../../../../tests/mocks/server";
 
 const nextSelector = () => screen.getByTestId("Paginator__continue-button");
 const backSelector = () => screen.getByTestId("Paginator__back-button");
